feat: move Settings button to a separate component
