+++ app/(tabs)/Home.tsx
@@ -8,7 +8,6 @@ import {
   useWindowDimensions,
   TouchableOpacity,
 } from 'react-native'
-import {Settingsicon} from '@/components/ui/Bottomnav/Icons'
 import ButtonImage from '@/components/ui/ButtonImage'
 import Animated, {
   FadeInDown,
@@ -19,19 +18,7 @@ import Animated, {
   withTiming,
   withDelay,
 } from 'react-native-reanimated'
-
-const AnimatedTouchableOpacity =
-  Animated.createAnimatedComponent(TouchableOpacity)
-
-const Settingsbutton = () => {
-  return (
-    <AnimatedTouchableOpacity
-      entering={FadeInDown.delay(300).duration(500)}
-      className='flex items-center justify-center mt-10 h-12 w-12 rounded-xl bg-[#518893] border-4 border-[#6CB0B4]'>
-      <Settingsicon size={22} />
-    </AnimatedTouchableOpacity>
-  )
-}
+import Settingsbutton from '@/components/ui/Settingsbutton'
 
 export default function Home() {
   const {width} = useWindowDimensions()
